Optimized launch_op compilation
